pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "samus-agent"
version = "0.1.0"
description = "Samus-inspired autonomous agent using Claude Sonnet 4"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Samus Agent Team"},
]
keywords = ["ai", "agent", "autonomous", "mcp", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "GitPython>=3.1.0",
    "PyGithub>=2.1.0",
    "openai>=1.35.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
samus = "samus.cli:main"

[project.urls]
Homepage = "https://github.com/samus-agent/samus-agent"
Repository = "https://github.com/samus-agent/samus-agent"
Issues = "https://github.com/samus-agent/samus-agent/issues"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["samus"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true